% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcamR.R
\name{webcamR-shiny}
\alias{webcamR-shiny}
\alias{webcamROutput}
\alias{renderWebcamR}
\alias{webcamR_html}
\title{Shiny bindings for webcamR}
\usage{
webcamROutput(outputId, width = "100\%\%", height = "400px")

renderWebcamR(expr, env = parent.frame(), quoted = FALSE)

webcamR_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a webcamR}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using webcamR within Shiny
applications and interactive Rmd documents.
}
